requirements.txt
# gRPC Dependencies
grpcio==1.62.0
grpcio-tools==1.62.0
grpcio-testing==1.62.0
protobuf==4.25.3

# Testing & Mocking
pytest==7.4.4
pytest-asyncio==0.23.5
pytest-grpc==0.8.0
pytest-mock==3.12.0
pytest-cov==4.1.0

# Async Support
aiohttp==3.9.3
async-timeout==4.0.3
asyncio==3.4.3

# Logging & Monitoring
structlog==24.1.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
prometheus-client==0.19.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.1
graphql-core==3.2.3
xmltodict==0.13.0
jwt==1.3.1
markdown2==2.4.12
jinja2==3.1.3